\name{Personality-package}
\alias{Personality-package}
\alias{Personality}
\docType{package}
\title{
    Test somenone's personality
}
\description{
    You can use this function to do a simple test about your personality and have some fun! Just type your name and answer the questions.
}
\details{

The DESCRIPTION file:
   This packsge is made to test someone's personality, just type your name and answer the questions and you will get the result! For example, if your name is Jack, then use the function this way: "Personality_test("Jack")".
}
\author{
\packageAuthor{Personality}

Maintainer: \packageMaintainer{Personality}
}
\references{
    Nothing
}

\keyword{ package }
\seealso{
    Nothing
}
\examples{
    For example, if your name is Jack, then use the function this way: "Personality_test("Jack")".
}
